[MISC] Move enumeration_names CPO into it's own header.
