+++ test/snippet/validators_input_file_ext_from_file.cpp
@@ -1,3 +1,5 @@
+#include <iostream>
+
 #include <sharg/validators.hpp>
 
 int main()
